+++ app/Plugin/Installer.php
@@ -63,7 +63,7 @@ class Installer {
 			// loop through the blogs.
 			foreach ( Helper::get_blogs() as $blog_id ) {
 				// switch to the blog.
-				switch_to_blog( $blog_id->blog_id );
+				switch_to_blog( $blog_id['blog_id'] );
 
 				// run tasks for activation in this single blog.
 				$this->activation_tasks();
@@ -112,7 +112,7 @@ class Installer {
 		// set marker to refresh permalinks.
 		update_option( 'personio_integration_update_slugs', 1 );
 
-		// show success message on cli.
+		// show a success message on cli.
 		Helper::is_cli() ? \WP_CLI::success( 'Personio Integration Light activated. Thank you for using our plugin :-)' ) : false;
 	}
 }