+++ app/PersonioIntegration/Positions.php
@@ -204,7 +204,7 @@ class Positions {
 		$this->results = new WP_Query( $query );
 
 		// add log for this query if debug is enabled.
-		if ( 1 === absint( get_option( 'personioIntegration_debug', 0 ) ) ) {
+		if ( 1 === absint( get_option( 'personioIntegrationQueryDebug', 0 ) ) ) {
 			Log::get_instance()->add( __( 'Query-Debug', 'personio-integration-light' ) . ': <code>' . wp_json_encode( $query ) . '</code><br><br>' . __( 'Result', 'personio-integration-light' ) . ': <code>' . wp_json_encode( $this->get_results() ) . '</code><br><br>' . __( 'Used URL', 'personio-integration-light' ) . ': ' . esc_url( Helper::get_current_url() ), 'info', 'system' );
 		}