+++ app/Plugin/Init.php
@@ -199,7 +199,7 @@ class Init {
 
 		// add our custom links.
 		$row_meta = array(
-			'support' => '<a href="' . esc_url( Helper::get_plugin_support_url() ) . '" target="_blank" title="' . esc_html__( 'Support Forum', 'personio-integration-light' ) . '">' . esc_html__( 'Support Forum', 'personio-integration-light' ) . '</a>',
+			'support' => '<a href="' . esc_url( Helper::get_plugin_support_url() ) . '" target="_blank" title="' . esc_attr__( 'Support Forum', 'personio-integration-light' ) . '">' . esc_html__( 'Support Forum', 'personio-integration-light' ) . '</a>',
 		);
 
 		/**