+++ app/Plugin/Settings.php
@@ -595,15 +595,28 @@ class Settings {
 		$setting->set_field( $field );
 
 		// add setting.
-		$setting = $settings_obj->add_setting( 'personioIntegration_debug' );
-		$setting->set_section( $advanced );
-		$setting->set_type( 'integer' );
-		$setting->set_default( 0 );
+		$debug_setting = $settings_obj->add_setting( 'personioIntegration_debug' );
+		$debug_setting->set_section( $advanced );
+		$debug_setting->set_type( 'integer' );
+		$debug_setting->set_default( 0 );
 		$field = new Checkbox();
 		$field->set_title( __( 'Debug-Mode', 'personio-integration-light' ) );
 		$field->set_readonly( ! Helper::is_personio_url_set() );
 		/* translators: %1$s will be replaced by a URL. */
 		$field->set_description( sprintf( __( 'When activated, the plugin logs many processes. This information can then be seen <a href="%1$s">in the log</a>. This helps to analyze any problems that may occur. At the same time, all open positions are retrieved in full at any time - it will not be checked whether anything has been changed in Personio. <strong>We do not recommend using this mode permanently in a productive system.</strong>', 'personio-integration-light' ), esc_url( Helper::get_settings_url( 'personioPositions', 'logs' ) ) ) );
+		$debug_setting->set_field( $field );
+
+		// add setting.
+		$setting = $settings_obj->add_setting( 'personioIntegrationQueryDebug' );
+		$setting->set_section( $advanced );
+		$setting->set_type( 'integer' );
+		$setting->set_default( 0 );
+		$field = new Checkbox();
+		$field->set_title( __( 'Enable debug of database queries', 'personio-integration-light' ) );
+		$field->set_description( __( 'If activated, the plugin logs all database queries for positions. This information can then be seen in the log.', 'personio-integration-light' ) );
+
+		$field->set_readonly( ! Helper::is_personio_url_set() );
+		$field->add_depend( $debug_setting, 1 );
 		$setting->set_field( $field );
 
 		// add import.
@@ -947,6 +960,7 @@ class Settings {
 
 		// forward user to the dashboard.
 		wp_safe_redirect( get_admin_url() );
+		exit;
 	}
 
 	/**