+++ changelog.md
@@ -11,6 +11,7 @@
 ### Changed
 
 - Log if Personio does not send any last timestamp for the request of open positions
+- Run import if no timestamp is returned from Personio and debug mode is enabled
 - Optimized output of import results on WP CLI
 - Set translatable text for a button in site health
 - Some more PHP Unit Tests